+++ application/controllers/Welcome.php
@@ -23,10 +23,29 @@ class Welcome extends Application
 		$authors = array ();
 		foreach ($source as $record)
 		{
-			$authors[] = array ('who' => $record['who'], 'mug' => $record['mug'], 'href' => $record['where']);
+			$authors[] = array ('who' => $record['who'], 'mug' => $record['mug'], 'href' => $record['where'], 'what' => $record['what']);
 		}
 		$this->data['authors'] = $authors;
 
+		$this->render();
+	}
+	public function random()
+	{
+		// this is the view we want shown
+		$this->data['pagebody'] = 'homepage';
+		$randId = rand(1,7);
+		// build the list of authors, to pass on to our view
+		$source = $this->quotes->all();
+		$authors = array ();
+		foreach ($source as $record)
+		{
+			if ($record['id'] == $randId){
+				$authors[] = array ('who' => $record['who'], 'mug' => $record['mug'], 'href' => $record['where'], 'what' => $record['what']);
+				break;
+			}
+			
+		}
+		$this->data['authors'] = $authors;
 		$this->render();
 	}