+++ changelog.md
@@ -4,3 +4,14 @@ Made by Harold Lee & Tommy Tran
 Style convention : Allman notation 
 Addition newest first 
 This project is the starting repository for COMP4711 (fall 2016) lab #2.
+
+
+Version 2.0.0
+- added random quotes 
+
+Version 1.0.1
+- added picture and quotes
+- fixed controller/Welcome
+
+Version 1.0.0
+- Initialized repository